python/requirements.txt
python-medium >= 1.1
click >= 8.1.7
python-env >= 1.0.0